+++ CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
-- Separate generated types from `directLine.ts` to `types.ts`
+- Refactor generated types from `directLine.ts` to new `types.ts`
 
 ## [0.13.0] - 2020-08-17