+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Bumped [`botframework-streaming@4.10.3`](https://npmjs.com/package/botframework-streaming), by [@stevengum](https://github.com/stevengum), in PR [#308](https://github.com/microsoft/BotFramework-DirectLineJS/pull/308)
+
 ## [0.13.0] - 2020-08-17
 
 ### Added